[#2608] clean up craptastic build system and update to python 2.4.4 ... this gets us the core python binary
git-svn-id: svn://localhost/svn/uclinux-dist/trunk@4802 3e2a5524-39e8-4cb9-84a9-c49da69cba3a

+++ user/python/Python/hypot.c
@@ -1,16 +1,16 @@
 /* hypot() replacement */
 
-#include "config.h"
+#include "pyconfig.h"
 #include "pyport.h"
 
-float hypot(float x, float y)
+double hypot(double x, double y)
 {
-	float yx;
+	double yx;
 
 	x = fabs(x);
 	y = fabs(y);
 	if (x < y) {
-		float temp = x;
+		double temp = x;
 		x = y;
 		y = temp;
 	}
